requirements.txt
wheel
torch==2.9.1
diffusers==0.33.0
nn-dataset>=2.2.6
nn-rag>=2.2.4
nn-dup>=2.2.4
bitsandbytes>=0.45.0
peft>=0.14.0
gguf>=0.10.0
trl
sentence-transformers>=3.3.1
faiss-cpu
overrides>=7.7.0
datasketch # to check code novelty

# Unsloth dependencies (official installation)
unsloth_zoo[base] @ git+https://github.com/unslothai/unsloth-zoo
unsloth[base] @ git+https://github.com/unslothai/unsloth
triton_kernels @ git+https://github.com/triton-lang/triton.git@05b2c186c1b6c9a08375389d5efe9cb4c401c075#subdirectory=python/triton_kernels
triton>=3.4.0

matplotlib
optimum
optimum[onnxruntime-gpu]
onnx>=1.16.0
onnxruntime-gpu>=1.18.0
